+++ src/components/content/patient-record-search/patient-record-search.tsx
@@ -9,7 +9,7 @@ import { ErrorAlert } from "@/components/core/alert";
 import { Title } from "@/components/core/ctw-box";
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { Loading } from "@/components/core/loading";
-import { DEFAULT_PAGE_SIZE, PaginationList } from "@/components/core/pagination/pagination-list";
+import { DEFAULT_PAGE_SIZE, ExpandList } from "@/components/core/pagination/expand-list";
 import { AnalyticsProvider } from "@/components/core/providers/analytics/analytics-provider";
 import { useAnalytics } from "@/components/core/providers/analytics/use-analytics";
 import { RenderIf } from "@/components/core/render-if";
@@ -60,7 +60,7 @@ function PatientRecordSearchComponent({ className, hideTitle = false }: PatientR
     setSearchValue("");
   }, [trackInteraction]);
 
-  const userHasSearched = !!data.query;
+  const userHasSearched = !!searchValue;
   const hasResults = data.results.length > 0;
 
   return (
@@ -109,13 +109,13 @@ function PatientRecordSearchComponent({ className, hideTitle = false }: PatientR
               </span>
             </RenderIf>
 
-            <RenderIf condition={userHasSearched && !hasResults}>
+            <RenderIf condition={userHasSearched && !hasResults && !isError}>
               <span className="ctw-text-1xl ctw-text-content-dark ctw-block ctw-text-left ctw-font-medium">
                 Search did not return any results.
               </span>
             </RenderIf>
 
-            <RenderIf condition={userHasSearched && isError}>
+            <RenderIf condition={isError}>
               <div className="ctw-w-full">
                 <ErrorAlert header="Error">There was an error running your search.</ErrorAlert>
               </div>
@@ -142,7 +142,7 @@ function PatientRecordSearchComponent({ className, hideTitle = false }: PatientR
                 ))}
               </FeedbackProvider>
               <div className="ctw-mt-5">
-                <PaginationList total={data.results.length} count={count} changeCount={setCount} />
+                <ExpandList total={data.results.length} count={count} changeCount={setCount} />
               </div>
             </RenderIf>
           </div>