+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # ctw-component-library
 
+## 1.65.16
+
+### Patch Changes
+
+- 007a1ee9: Add full pagination to ADT table.
+- 4a9d23e1: Add letterSpacing and fontWeight to iframeTheme Props
+
+## 1.65.15
+
+### Patch Changes
+
+- b3c0a641: Fix error state in search component.
+- aba449c6: Address various clinical note rendering issues
+
+## 1.65.14
+
+### Patch Changes
+
+- 056c8e8d: Fixed issue where clinical notes were not visible
+
 ## 1.65.13
 
 ### Patch Changes